requirements.txt
fastapi==0.115.12
uvicorn[standart]
SQLAlchemy==2.0.41
aiosqlite==0.21.0
alembic==1.15.2
httpx==0.28.1
pydantic-settings==2.9.1
passlib==1.7.4
PyJWT==2.10.1
pydantic[email]
python-multipart==0.0.20
